+++ tests/models/test_models.py
@@ -442,14 +442,25 @@ def test_trial_metadata_get_summaries(clean_db, monkeypatch):
         **METADATA,
         # deliberately override METADATA['protocol_identifier']
         "protocol_identifier": "tm1",
-        "participants": [{"samples": [1, 2]}, {"samples": [3]}],
+        "participants": [
+            {
+                "samples": [
+                    {"cimac_id": "1", "processed_sample_derivative": "Tumor DNA"},
+                    {"cimac_id": "2", "processed_sample_derivative": "not"},
+                ]
+            },
+            {
+                "samples": [
+                    {"cimac_id": "3", "processed_sample_derivative": "Tumor DNA"}
+                ]
+            },
+        ],
         "expected_assays": ["ihc", "olink"],
         "assays": {
             "atacseq": [{"records": records * 13}],
             "wes": [
-                {"records": records * 6},
-                {"records": records * 5},
-            ],  # 6 + 5 11 = 7 for wes + 4 for wes_tumor_only
+                {"records": [{"cimac_id": str(n)} for n in range(1, 4)]},
+            ],  # 2 for wes_tumor + 1 for wes_normal
             "rna": [{"records": records * 2}],
             "mif": [
                 {"records": records * 3},
@@ -467,11 +478,11 @@ def test_trial_metadata_get_summaries(clean_db, monkeypatch):
             "atacseq_analysis": [{"records": records * 12}],
             "wes_analysis": {
                 "pair_runs": [
-                    # 7 here for wes_assay: t0/1/2, n0/1/2/3
                     {
                         "tumor": {"cimac_id": "t0"},
                         "normal": {"cimac_id": "n0"},
                     },  # no analysis data
+                    # 3 for wes_analysis
                     {
                         "tumor": {"cimac_id": "t1"},
                         "normal": {"cimac_id": "n1"},
@@ -492,7 +503,7 @@ def test_trial_metadata_get_summaries(clean_db, monkeypatch):
                 "excluded_samples": records * 2,
             },
             "wes_tumor_only_analysis": {
-                "runs": records * 4,  # need 4
+                "runs": records * 4,
                 # these are excluded, so not adding fake assay data
                 "excluded_samples": records * 3,
             },
@@ -583,8 +594,8 @@ def test_trial_metadata_get_summaries(clean_db, monkeypatch):
                 "tcr_analysis": 6.0,
                 "wes_analysis": 0.0,
                 "wes_tumor_only_analysis": 0.0,
-                "wes": 0.0,
-                "wes_tumor_only": 0.0,
+                "wes_tumor": 0.0,
+                "wes_normal": 0.0,
                 "excluded_samples": {
                     "tcr_analysis": records * 4,
                     "rna_level1_analysis": records * 2,
@@ -610,10 +621,10 @@ def test_trial_metadata_get_summaries(clean_db, monkeypatch):
                 "mif": 5.0,
                 "rna_level1_analysis": 0.0,
                 "tcr_analysis": 0.0,
-                "wes_analysis": 5.0,
+                "wes_analysis": 3.0,
                 "wes_tumor_only_analysis": 4.0,
-                "wes": 7.0,
-                "wes_tumor_only": 4.0,
+                "wes_normal": 1.0,
+                "wes_tumor": 2.0,
                 "excluded_samples": {
                     "wes_analysis": records * 2,
                     "wes_tumor_only_analysis": records * 3,