+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This Changelog tracks changes to this project. The notes below include a summary
 - `fixed` for any bug fixes.
 - `security` in case of vulnerabilities.
 
+## Version `0.26.11` - 19 May 2022
+
+- `changed` order of user and url in single file download
+- `changed` WES counting to reflect new system
+- `added` new downloadable_files endpoint to return all facets grouped by assay
+
 ## 6 May 2022
 
 - `removed` `security` HSTS that was broken